django-rest/requirements.txt
asgiref==3.5.0
boto3==1.20.41
botocore==1.23.41
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
jmespath==0.10.0
psycopg2-binary==2.9.3
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
